moved functions to utils folder
